# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TripMatch API - Group travel planning with swipe-based matching
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile and preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /swipes:
    post:
      operationId: recordSwipe
      tags: [swipes]
      summary: Record a swipe on a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordSwipeBody"
      responses:
        "201":
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResult"
  /swipes/group/{groupId}:
    get:
      operationId: listGroupSwipes
      tags: [swipes]
      summary: List all swipes for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of swipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Swipe"
  /swipes/user/{groupId}:
    get:
      operationId: listUserSwipes
      tags: [swipes]
      summary: List current user's swipes in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's swipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Swipe"
...
components:
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        displayName:
          type: string
        profileImageUrl:
          type: ["string", "null"]
        preferences:
          $ref: "#/components/schemas/UserPreferences"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - displayName
        - createdAt
    UserPreferences:
      type: object
      properties:
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        travelTypes:
          type: array
          items:
            type: string
        climate:
          type: ["string", "null"]
        activityLevel:
          type: ["string", "null"]
      required: []
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        preferences:
          $ref: "#/components/schemas/UserPreferences"
...
    RecordSwipeBody:
      type: object
      properties:
        groupId:
          type: integer
        destinationId:
          type: integer
        value:
          type: integer
          description: "-1=dislike, 1=like, 2=superlike"
      required:
        - groupId
        - destinationId
        - value
    Swipe:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        groupId:
          type: integer
        destinationId:
          type: integer
        value:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - groupId
        - destinationId
        - value
        - createdAt
    SwipeResult:
      type: object
      properties:
        swipe:
          $ref: "#/components/schemas/Swipe"
        updatedScore:
          $ref: "#/components/schemas/DestinationScore"
      required:
        - swipe
        - updatedScore
    DestinationScore:
      type: object
      properties:
        destinationId:
          type: integer
        destinationName:
          type: string
        country:
          type: string
        imageUrl:
          type: string
        groupScore:
          type: number
        matchPercentage:
          type: number
        consensusLevel:
          type: string
          enum: [high, medium, low]
        likeCount:
          type: integer
        dislikeCount:
          type: integer
        superlikeCount:
          type: integer
        totalVotes:
          type: integer
      required:
        - destinationId
        - destinationName
        - country
        - imageUrl
        - groupScore
        - matchPercentage
        - consensusLevel
        - likeCount
        - dislikeCount
        - superlikeCount
        - totalVotes
...