# lib/api-zod/src/generated/types/swipe.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripMatch API - Group travel planning with swipe-based matching
 * OpenAPI spec version: 0.1.0
 */

export interface Swipe {
  id: number;
  userId: string;
  groupId: number;
  destinationId: number;
  value: number;
  createdAt: Date;
}