# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripMatch API - Group travel planning with swipe-based matching
 * OpenAPI spec version: 0.1.0
 */
import type { UserPreferences } from "./userPreferences";

export interface User {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  displayName: string;
  /** @nullable */
  profileImageUrl?: string | null;
  preferences?: UserPreferences;
  createdAt: Date;
}